fix trigger ordering
